Scripts for ea. data model to faciliate big models
