+++ .vscode/settings.json
@@ -131,6 +131,7 @@
     "testpypi",
     "timegrid",
     "uids",
+    "ulid",
     "unfinish",
     "unitsofmeasure",
     "untracked",
