+++ cannlytics/models/cannabis_license.py
@@ -1,0 +1,25 @@
+"""
+Cannabis License Model | Cannlytics
+Copyright (c) 2021 Cannlytics and Cannlytics Contributors
+
+Author: Keegan Skeate <keegan@cannlytics.com>  
+Created: 11/5/2021  
+Updated: 11/5/2021  
+
+Cannabis license data model.
+"""
+# Standard imports.
+from dataclasses import dataclass
+
+# Internal imports.
+from .base import Model
+
+
+@dataclass
+class CannabisLicense(Model):
+    """A state-issued cannabis license."""
+    _collection = 'organizations/%s/licenses'
+    active_date: str = ''
+    expiration_date: str = ''
+    license_number: str = ''
+    license_type: str = ''
